+++ client/src/components/IllustGroup.tsx
@@ -1,9 +1,9 @@
-import React, { useMemo, useRef, useState } from "react";
+import React, { useMemo, useState } from "react";
 import { useInView } from "react-intersection-observer";
 import styled from "styled-components";
 import { blurRange, defaultBoxShadow, getKeyColor } from "@/const/styles";
-import { getImageEndpoint, Tweet, TweetToTag, Work } from "@/utils/api";
-import { DisplayOptions } from "@/utils/utils";
+import { getImageEndpoint, Tweet } from "@/utils/api";
+import { getImagePixel, loadImage, Color, DisplayOptions } from "@/utils/utils";
 
 const Wrapper = styled.div`
   padding: 20px 0;
@@ -53,15 +53,27 @@ const DivImage = styled.div<{ src: string; blur: boolean }>`
 `;
 
 interface IllustGroupProps {
+  isShiftKeyPressed: boolean;
+  pixelColor: number[] | undefined;
+  mousePoint: [number, number] | undefined;
   selectedTweetIds: string[];
   tweets: Tweet[];
   displayOptions: DisplayOptions;
+  setPixelColor: (value: Color | undefined) => void;
+  setMousePoint: (value: [number, number] | undefined) => void;
+  setSelectedTweetIds: (value: string[]) => void;
 }
 
 const IllustGroup = ({
+  isShiftKeyPressed,
+  pixelColor,
+  mousePoint,
   selectedTweetIds,
   tweets,
   displayOptions,
+  setPixelColor,
+  setMousePoint,
+  setSelectedTweetIds,
 }: IllustGroupProps) => {
   const [aspectRatios, setAspectRatios] = useState<number[]>([]);
 
@@ -100,6 +112,55 @@ const IllustGroup = ({
     return tweetsPerColumn;
   }, [tweets, displayOptions]);
 
+  const onClickTweet = async (e: React.MouseEvent, id: string) => {
+    // color picker
+    if (e.ctrlKey || e.metaKey) {
+      const rect = e.currentTarget.getBoundingClientRect();
+
+      // get color
+      const img = new Image();
+      img.src = getImageEndpoint(id, 0);
+      await loadImage(img);
+
+      let x = e.clientX - rect.x;
+      let y = e.clientY - rect.y;
+      if (displayOptions.isSquare) {
+        if (img.naturalWidth > img.naturalHeight) {
+          x += ((img.naturalWidth / img.naturalHeight - 1.0) * rect.height) / 2;
+        } else {
+          y += ((img.naturalHeight / img.naturalWidth - 1.0) * rect.width) / 2;
+        }
+      }
+
+      const pixel = await getImagePixel(
+        img,
+        x / rect.width,
+        y / ((img.height / img.width) * rect.width),
+        rect.width
+      );
+
+      setPixelColor(pixel);
+      setMousePoint([e.clientX, e.clientY]);
+      if (selectedTweetIds.length === 0) {
+        setSelectedTweetIds([id]);
+      }
+    } else {
+      if (!pixelColor || !mousePoint) {
+        setSelectedTweetIds(
+          isShiftKeyPressed
+            ? selectedTweetIds.includes(id)
+              ? selectedTweetIds.filter((inId) => inId !== id)
+              : [...selectedTweetIds, id]
+            : selectedTweetIds.includes(id)
+            ? []
+            : [id]
+        );
+      }
+      setPixelColor(undefined);
+      setMousePoint(undefined);
+    }
+  };
+
   return (
     <Wrapper ref={wrapperRef}>
       {tweetRows.map((row, i) => (
@@ -114,7 +175,7 @@ const IllustGroup = ({
                 <Illust
                   aspectRatio={displayOptions.isSquare ? 1 : getRatio(tweet)}
                   selected={selectedTweetIds.includes(tweet.id)}
-                  //onClick={(e) => onClickTweet(e, tweet.id)}
+                  onClick={(e) => onClickTweet(e, tweet.id)}
                   key={tweet.id}
                 >
                   <DivImage
